Move color widgets to ui
