+++ src/app/GUI/ColorWidgets/colorsettingswidget.cpp
@@ -30,10 +30,10 @@
 #include <QMenu>
 #include "GUI/ColorWidgets/colorpickingwidget.h"
 #include "colorhelpers.h"
-#include "colorlabel.h"
+#include "widgets/colorlabel.h"
 #include "GUI/global.h"
 #include "widgets/actionbutton.h"
-#include "GUI/ColorWidgets/savedcolorswidget.h"
+#include "widgets/savedcolorswidget.h"
 #include "appsupport.h"
 
 #include <QShortcut>
