+++ src/ui/widgets/colorlabel.cpp
