+++ src/ui/widgets/colorvaluerect.h
@@ -25,17 +25,23 @@
 
 #ifndef COLORVALUERECT_H
 #define COLORVALUERECT_H
-#include "colorwidget.h"
+
+#include "ui_global.h"
+
+#include "widgets/colorwidget.h"
 #include "widgets/colorwidgetshaders.h"
 
-class ColorValueRect : public ColorWidget {
+class UI_EXPORT ColorValueRect : public ColorWidget
+{
     Q_OBJECT
+
 public:
     ColorValueRect(const ColorProgram &program,
                    QWidget * const parent = nullptr);
 
     void setDisplayedValue(const qreal val_t);
     qreal value() const { return mVal; }
+
 private:
     void paintGL();
     void mouseMoveEvent(QMouseEvent *e);
@@ -46,10 +52,12 @@ private:
     void mouseInteraction(const int x_t);
     void setColorParameterFromVal();
     void setValueAndEmitValueChanged(const qreal valT);
+
 signals:
     void valueChanged(qreal);
     void editingFinished(qreal);
     void editingStarted(qreal);
+
 private:
     const ColorProgram& mGLProgram;
     qreal mVal = 0;
