+++ src/ui/widgets/colorwidget.cpp
