+++ src/ui/widgets/savedcolorbutton.cpp
