+++ src/ui/widgets/savedcolorswidget.cpp
@@ -24,7 +24,7 @@
 // Fork of enve - Copyright (C) 2016-2020 Maurycy Liebner
 
 #include "savedcolorswidget.h"
-#include "savedcolorbutton.h"
+#include "widgets/savedcolorbutton.h"
 #include <QAction>
 #include <QMenu>
 #include "colorhelpers.h"
